requirements.txt
emoji==2.2.0
pyTelegramBotAPI==4.10.0
requests==2.28.2
python-dotenv~=1.0.0